ObjectMappings now overridden when index is reused
The OD entries responsible for PDO mappings may be reused,
this is now handled.

The MN has to set ObjectMappings to zero prior to changing PDO Mappings
We don't require tht in the dissector however.
A MN might set it to zero and keep it at that, this would result in us
using stale object mappings, which could lead to incorrect dissection.
But the user can always check the hex output. So compared with not dissecting
at all, it's an acceptable tradeoff.