LibXML now initialized once only
Previously was reinitialized for each capture file.
This comes with the drawback that we can't call xdd_free, because there's no place for it AFAIK,
But I am not even sure, I should. What if the GTK GUI makes use libxml after we free it? I guess it's better to leak it?!